restructure with run_suite
