+++ test/registered/stress/test_stress_qwen3_235b.py
@@ -1,19 +1,23 @@
-"""Stress test for GLM-4.6 model."""
+"""Stress test for Qwen3-235B model."""
 
 import os
 import unittest
 
-from stress_test_utils import StressTestRunner
+from stress_utils import StressTestRunner
 
+from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST
 
-MODEL_PATH = "zai-org/GLM-4.6"
+MODEL_PATH = "Qwen/Qwen3-235B-A22B-Instruct-2507"
 RANDOM_INPUT_LEN = 4096
 RANDOM_OUTPUT_LEN = 512
-OUTPUT_FILE = "stress_test_glm_4_6.jsonl"
+OUTPUT_FILE = "stress_test_qwen3_235b.jsonl"
+
+# Register for CI - estimated 45 minutes
+register_cuda_ci(est_time=2700, suite="stress")
 
 
-class TestStressGLM46(unittest.TestCase):
+class TestStressQwen3235B(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.model = MODEL_PATH
@@ -22,13 +26,13 @@ class TestStressGLM46(unittest.TestCase):
         cls.duration_minutes = int(os.environ.get("DURATION_MINUTES", "45"))
 
         cls.runner = StressTestRunner(
-            test_name="GLM-4.6 Stress Test",
+            test_name="Qwen3-235B Stress Test",
             base_url=cls.base_url,
             num_prompts=cls.num_prompts,
             duration_minutes=cls.duration_minutes,
         )
 
-    def test_stress_glm_4_6(self):
+    def test_stress_qwen3_235b(self):
         try:
             success = self.runner.run_stress_test_for_model(
                 model_path=self.model,
